Documentation updated.

+++ public/offline/models/synchronization.js
@@ -1,10 +1,11 @@
-/**
- * @tag models, home
- * Wraps backend synchronization services.
- */
 $.Model.extend('Synchronization',
 /* @Static */
 {
+    /**
+     * Starts sync process.
+     * @param {Function} success - a callback function that returns wrapped synchronization objects.
+     * @param {Function} error - a callback function for an error.
+     */
     start: function(success, error) {
         var obj = new Synchronization()
         obj.sync_remote_feedback_table()
@@ -19,11 +20,12 @@ $.Model.extend('Synchronization',
         localStorage['last_sync'] = (new Date()).getTime()
         return success
     },
+
     /**
      * Retrieves synchronizations data from your backend services.
-     * @param {Object} params params that might refine your results.
-     * @param {Function} success a callback function that returns wrapped synchronization objects.
-     * @param {Function} error a callback function for an error in the ajax request.
+     * @param {Array} params - params that might refine your results.
+     * @param {Function} success - a callback function that returns wrapped synchronization objects.
+     * @param {Function} error - a callback function for an error in the ajax request.
      */
     blog : function(params, success, error){
         $.ajax({
@@ -34,14 +36,15 @@ $.Model.extend('Synchronization',
             success: this.callback(['wrapMany',success]),
             error: error,
             async: false,
-            fixture: false //calculates the fixture path from the url and type.
+            fixture: false
         })
     },
+    
     /**
      * Retrieves synchronizations data from your backend services.
-     * @param {Object} params params that might refine your results.
-     * @param {Function} success a callback function that returns wrapped synchronization objects.
-     * @param {Function} error a callback function for an error in the ajax request.
+     * @param {Array} params - params that might refine your results.
+     * @param {Function} success - a callback function that returns wrapped synchronization objects.
+     * @param {Function} error - a callback function for an error in the ajax request.
      */
     categories : function(params, success, error){
         $.ajax({
@@ -52,14 +55,15 @@ $.Model.extend('Synchronization',
             success: this.callback(['wrapMany',success]),
             error: error,
             async: false,
-            fixture: false //calculates the fixture path from the url and type.
+            fixture: false
         })
     },
+
     /**
      * Retrieves synchronizations data from your backend services.
-     * @param {Object} params params that might refine your results.
-     * @param {Function} success a callback function that returns wrapped synchronization objects.
-     * @param {Function} error a callback function for an error in the ajax request.
+     * @param {Array} params - params that might refine your results.
+     * @param {Function} success - a callback function that returns wrapped synchronization objects.
+     * @param {Function} error - a callback function for an error in the ajax request.
      */
     categorizations : function(params, success, error){
         $.ajax({
@@ -70,14 +74,15 @@ $.Model.extend('Synchronization',
             success: this.callback(['wrapMany',success]),
             error: error,
             async: false,
-            fixture: false //calculates the fixture path from the url and type.
+            fixture: false
         })
     },
+
     /**
      * Retrieves synchronizations data from your backend services.
-     * @param {Object} params params that might refine your results.
-     * @param {Function} success a callback function that returns wrapped synchronization objects.
-     * @param {Function} error a callback function for an error in the ajax request.
+     * @param {Array} params - params that might refine your results.
+     * @param {Function} success - a callback function that returns wrapped synchronization objects.
+     * @param {Function} error - a callback function for an error in the ajax request.
      */
     contents : function(params, success, error){
         $.ajax({
@@ -88,14 +93,15 @@ $.Model.extend('Synchronization',
             success: this.callback(['wrapMany',success]),
             error: error,
             async: false,
-            fixture: false //calculates the fixture path from the url and type.
+            fixture: false
         })
     },
+
     /**
      * Retrieves synchronizations data from your backend services.
-     * @param {Object} params params that might refine your results.
-     * @param {Function} success a callback function that returns wrapped synchronization objects.
-     * @param {Function} error a callback function for an error in the ajax request.
+     * @param {Array} params - params that might refine your results.
+     * @param {Function} success - a callback function that returns wrapped synchronization objects.
+     * @param {Function} error - a callback function for an error in the ajax request.
      */
     feedback : function(params, success, error){
         $.ajax({
@@ -106,14 +112,15 @@ $.Model.extend('Synchronization',
             success: this.callback(['wrapMany',success]),
             error: error,
             async: false,
-            fixture: false //calculates the fixture path from the url and type.
+            fixture: false
         })
     },
+
     /**
      * Retrieves synchronizations data from your backend services.
-     * @param {Object} params params that might refine your results.
-     * @param {Function} success a callback function that returns wrapped synchronization objects.
-     * @param {Function} error a callback function for an error in the ajax request.
+     * @param {Array} params - params that might refine your results.
+     * @param {Function} success - a callback function that returns wrapped synchronization objects.
+     * @param {Function} error - a callback function for an error in the ajax request.
      */
     sidebars : function(params, success, error){
         $.ajax({
@@ -124,14 +131,15 @@ $.Model.extend('Synchronization',
             success: this.callback(['wrapMany',success]),
             error: error,
             async: false,
-            fixture: false //calculates the fixture path from the url and type.
+            fixture: false
         })
     },
+
     /**
      * Retrieves synchronizations data from your backend services.
-     * @param {Object} params params that might refine your results.
-     * @param {Function} success a callback function that returns wrapped synchronization objects.
-     * @param {Function} error a callback function for an error in the ajax request.
+     * @param {Array} params - params that might refine your results.
+     * @param {Function} success - a callback function that returns wrapped synchronization objects.
+     * @param {Function} error - a callback function for an error in the ajax request.
      */
     tags : function(params, success, error){
         $.ajax({
@@ -142,14 +150,15 @@ $.Model.extend('Synchronization',
             success: this.callback(['wrapMany',success]),
             error: error,
             async: false,
-            fixture: false //calculates the fixture path from the url and type.
+            fixture: false
         })
     },
+
     /**
      * Retrieves synchronizations data from your backend services.
-     * @param {Object} params params that might refine your results.
-     * @param {Function} success a callback function that returns wrapped synchronization objects.
-     * @param {Function} error a callback function for an error in the ajax request.
+     * @param {Array} params - params that might refine your results.
+     * @param {Function} success - a callback function that returns wrapped synchronization objects.
+     * @param {Function} error - a callback function for an error in the ajax request.
      */
     articles_tags : function(params, success, error){
         $.ajax({
@@ -160,14 +169,15 @@ $.Model.extend('Synchronization',
             success: this.callback(['wrapMany',success]),
             error: error,
             async: false,
-            fixture: false //calculates the fixture path from the url and type.
+            fixture: false
         })
     },
+
     /**
-     * Creates a synchronization.
-     * @param {Object} attrs A synchronization's attributes.
-     * @param {Function} success a callback function that indicates a successful create.  The data that comes back must have an ID property.
-     * @param {Function} error a callback that should be called with an object of errors.
+     * Retrieves synchronizations data from your backend services.
+     * @param {Array} params - params that might refine your results.
+     * @param {Function} success - a callback function that returns wrapped synchronization objects.
+     * @param {Function} error - a callback function for an error in the ajax request.
      */
     update_feedback : function(attrs, success, error){
         $.ajax({
@@ -183,6 +193,7 @@ $.Model.extend('Synchronization',
 },
 /* @Prototype */
 {
+    //sync feedback table
     sync_remote_feedback_table: function() {
         db.transaction(function(tx) {
             tx.executeSql("SELECT * FROM sync WHERE table_name = ?",['feedback'],function(tx,rs){
@@ -192,8 +203,8 @@ $.Model.extend('Synchronization',
                     tx.executeSql("SELECT author,body,email,article_id,url FROM feedback WHERE id = ?",[row.row_id],
                         function(tx,rs){
                             var row = rs.rows.item(0)
-                            Synchronization.update_feedback("author="+row.author+"&body="+row.body+"&email="+row.email+"&url="+row.url+"&article_id="+row.article_id+"")
-                            tx.executeSql("DELETE FROM sync WHERE id = ?",[id])
+                            Synchronization.update_feedback("author="+row.author+"&body="+row.body+"&email="+row.email+"&url="+row.url+"&article_id="+row.article_id+"",tx.executeSql("DELETE FROM sync WHERE id = ?",[id]))
+//                            tx.executeSql("DELETE FROM sync WHERE id = ?",[id])
                         })
                 }
             },function(tx,error){
@@ -202,6 +213,7 @@ $.Model.extend('Synchronization',
             )
         })
     },
+
     //sync blog table
     sync_blog_table: function(result) {
         if(result.length > 0) {
@@ -224,6 +236,7 @@ $.Model.extend('Synchronization',
             })
         }
     },
+
     //sync articles_tags table
     sync_articles_tags_table: function(result) {
         if(result.length > 0) {
@@ -241,6 +254,7 @@ $.Model.extend('Synchronization',
             })
         }
     },
+
     //sync table categories
     sync_categories_table: function(result) {
         if(result.length > 0) {
@@ -258,6 +272,7 @@ $.Model.extend('Synchronization',
             })
         }
     },
+
     //sync table categorizations
     sync_categorizations_table: function(result) {
         if(result.length > 0) {
@@ -275,6 +290,7 @@ $.Model.extend('Synchronization',
             })
         }
     },
+
     //sync table sidebars
     sync_table_sidebars: function(result) {
         if(result.length > 0) {
@@ -293,6 +309,7 @@ $.Model.extend('Synchronization',
             })
         }
     },
+
     //sync table tags
     sync_table_tags: function(result) {
         if(result.length > 0) {
@@ -310,6 +327,7 @@ $.Model.extend('Synchronization',
             })
         }
     },
+
     //sync table contents
     sync_table_contents: function(result) {
         if(result.length > 0) {
@@ -334,6 +352,7 @@ $.Model.extend('Synchronization',
             })
         }
     },
+    
     //sync table feedback
     sync_table_feedback: function(result) {
         if(result.length > 0) {
